detect harness configuration (#1367)

+++ tools/harness-automation/autothreadharness/harness_case.py
@@ -27,6 +27,7 @@
 # POSSIBILITY OF SUCH DAMAGE.
 #
 
+import ConfigParser
 import json
 import logging
 import os
@@ -223,7 +224,12 @@ class HarnessCase(unittest.TestCase):
         time.sleep(1)
         self._hc.start()
         time.sleep(2)
-        os.system('taskkill /t /f /im chrome.exe')
+
+        harness_config = ConfigParser.ConfigParser()
+        harness_config.read('%s\\Config\\Configuration.ini' % settings.HARNESS_HOME)
+        if harness_config.has_option('THREAD_HARNESS_CONFIG', 'BrowserAutoNavigate') and \
+                harness_config.getboolean('THREAD_HARNESS_CONFIG', 'BrowserAutoNavigate'):
+            os.system('taskkill /t /f /im chrome.exe')
 
     def _destroy_harness(self):
         """Stop harness backend service
